+++ braindecode/models/util.py
@@ -132,6 +132,7 @@ models_mandatory_parameters: list[
             "chs_info": [{"ch_name": f"E{i + 1}", "kind": "eeg"} for i in range(19)],
         },
     ),
+    ("CBraMod", ["n_outputs"], None),
 ]
 
 ################################################################